some gui-projects
found it here: https://groups.google.com/forum/#!topic/emscripten-discuss/pcDZ18bm-TI
think it's worth to mention it although this list has a link to the oryol git-repo, it haven't one to the oryol-samples repo which has the guiDemos. In my opinion it's confusing to open two gui sectors one in projects and one in examples so i decided to make one GUIs section in projects which contains the sample gui-projects,too.